+++ CHANGES.md
@@ -87,6 +87,9 @@ support compositional verification using techniques similar to SAW.
 * Add new SAWScript MIR commands `mir_elem_value` and `mir_elem_ref` for
   indexing into arrays and references to arrays.
 
+* Add a new SAWScript MIR command `mir_vec_of` for creating Rust `Vec` values in
+  specifications.
+
 * Support writing MIR specifications involving raw pointers. This is done with
   the new SAWScript commands `mir_raw_ptr_const`, `mir_raw_ptr_mut`,
   `mir_alloc_raw_ptr_const`, and `mir_alloc_raw_ptr_mut`, which are similar to